Add a location endpoint to capture Overland data
[Overland](https://github.com/aaronpk/Overland-iOS) is an iOS app
used to capture location data from an iOS device. This data is then
posted to a customizable URL at a customizable interval. The
[Overland API](https://github.com/aaronpk/Overland-iOS#api)
describes the contract for this endpoint.

This commit introduces a new `/locations` endpoint to capture posted
data. A config value is used to store the pre-shared key for
authorization. This commit merely responds with success and throws away
the posted data.